Se dividieron los reducers actions y types entre dependientres y usuarios, corregi errores y asi
