+++ src/components/Dependientes/EditarDependiente.js
@@ -2,13 +2,13 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import {Redirect} from 'react-router-dom'
 import {Row, Input, Button, Preloader} from 'react-materialize';
-import * as UsuariosActions from '../../actions/UsuariosActions';
+import * as DependientesActions from '../../actions/DependientesActions';
 
 import {
 	 DEPENDENCIAEDITAR_NOMBRE,
 	 DEPENDENCIAEDITAR_DEPENDIENTE,
 	 DEPENDENCIAEDITAR_EDAD
-}from '../../types/UsuariosTypes';
+}from '../../types/DependientesTypes';
 
 class EditarDependiente extends Component {
 
@@ -88,9 +88,9 @@ class EditarDependiente extends Component {
 }
 
 
-const mapStateToProps =({ UsuariosReducers }) =>{
+const mapStateToProps =({ DependientesReducers }) =>{
 	
-	return UsuariosReducers;
+	return DependientesReducers;
 }
 
-export default connect(mapStateToProps, UsuariosActions)(EditarDependiente);
+export default connect(mapStateToProps, DependientesActions)(EditarDependiente);
