+++ src/reducers/index.js
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux';
 import UsuariosReducers from './UsuariosReducers';
+import DependientesReducers from './DependientesReducers';
 
 export default combineReducers({
-	UsuariosReducers
+	UsuariosReducers,
+	DependientesReducers
 });
