+++ src/types/DependientesTypes.js
@@ -1,0 +1,11 @@
+export const REINICIAR ='usuarios_reiniciar';
+export const PRELOADER ='preloader';
+export const TRUE = 'true';
+export const FALSE = 'false';
+export const  DEPENDENCIAEDITAR_NOMBRE ='DEPENDENCIAEDITAR_NOMBRE';
+export const  DEPENDENCIAEDITAR_EDAD ='DEPENDENCIAEDITAR_EDAD';
+export const  DEPENDENCIAEDITAR_DEPENDIENTE ='DEPENDENCIAEDITAR_DEPENDIENTE';
+export const dEXITO = 'DEPENDENCIAS_EXITO';
+export const TRAER = 'usuarios_traer';
+export const EXITO = 'usuarios_exito';
+export const FALLO = 'usuarios_fallo';
