+++ src/types/UsuariosTypes.js
@@ -9,7 +9,3 @@ export const REINICIAR ='usuarios_reiniciar';
 export const PRELOADER ='preloader';
 export const TRUE = 'true';
 export const FALSE = 'false';
-export const  DEPENDENCIAEDITAR_NOMBRE ='DEPENDENCIAEDITAR_NOMBRE';
-export const  DEPENDENCIAEDITAR_EDAD ='DEPENDENCIAEDITAR_EDAD';
-export const  DEPENDENCIAEDITAR_DEPENDIENTE ='DEPENDENCIAEDITAR_DEPENDIENTE';
-export const dEXITO = 'DEPENDENCIAS_EXITO';
